+++ libft/ft_putnbr_uns_base.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void	ft_putnbr_uns_base(unsigned int nb, char *charset)
+void	ft_putnbr_uns_base(unsigned long long int nb, char *charset)
 {
 	ft_putnbr_uns_base_fd(nb, 1, charset);
 }
