+++ libft/libft.h
@@ -64,15 +64,15 @@ int				ft_nbdigits_base(unsigned int nb, int base);
 void			ft_putchar_fd(char c, int fd);
 void			ft_putendl_fd(char *s, int fd);
 void			ft_putnbr_fd(int n, int fd);
-void			ft_putnbr_uns_fd(unsigned int nb, int fd);
+void			ft_putnbr_uns_fd(unsigned long long int nb, int fd);
 void			ft_putnbr_uns_base_fd(unsigned long long int u_n, int fd, char *charset);
 void			ft_putstr_fd(char *s, int fd);
 
 void			ft_putchar(char c);
 void			ft_putendl(char *s);
 void			ft_putnbr(int n);
-void			ft_putnbr_uns(unsigned int nb);
-void			ft_putnbr_uns_base(unsigned int nb, char *charset);
+void			ft_putnbr_uns(unsigned long long int nb);
+void			ft_putnbr_uns_base(unsigned long long int nb, char *charset);
 void			ft_putstr(char *s);
 
 void			ft_lstadd_back(t_list **alst, t_list *new);
